refactor damage system to use integers and improve projectile pooling #348

+++ Lords_Frontiers/Source/Lords_Frontiers/Public/Building/Building.h
@@ -34,7 +34,7 @@ public:
 
 	virtual ETeam Team() override;
 
-	virtual void TakeDamage( float damage ) override;
+	virtual void TakeDamage( int damage ) override;
 
 	virtual void OnSelected_Implementation() override;
